refactor(page): extract usePrintingMonitorManager (phase 3)
Move the printing-monitor domain (~2,950 lines / 193 statements: webcam
streaming + RTSP relay, device status polling, recent plates, target +
material picker, dashboard, debug bundles, printer reachability) out of
Home() into src/features/printing/usePrintingMonitorManager.ts.

State and all of its logic (memos, effects, callbacks) move together as
one cohesive manager hook — not a net-negative state-only split. The
boundary needed only 10 external deps, all defined before the call site,
so no deps-ref is required. Two trivial monitor-derived selectors shared
with the slice->print send flow (requiresRemoteMaterialSelectionForUpload,
isPreSliceTargetPicker) were folded into the hook and returned. The
send-flow glue stays in Home and reads the hook's ~206-name return via
destructure-same-names, so its call sites are unchanged.

Verbatim move via content-anchored AST slice-assemble (TS compiler API
for exact spans). page.tsx 12,432 -> 9,696 lines. tsc baseline-clean
(8 pre-existing __tests__ errors) and pnpm build green.
